requirements.txt

flask_jwt
flask_restful
flask_sqlalchemy